refactor: moves ethPriceChart async methods to utils folder and updates imports accordingly
